// client/src/api/generated/model/productModel.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * E-commerce Gateway API
 * Gateway REST API that proxies to Products, Categories, and Orders microservices via TCP
 * OpenAPI spec version: 1.0
 */
import type { ProductModelStatus } from './productModelStatus';
import type { CategoryModel } from './categoryModel';

export interface ProductModel {
  id: number;
  name: string;
  description: string;
  price: number;
  sellerName: string;
  imageUrl: string;
  status: ProductModelStatus;
  uploadDate: string;
  categories: CategoryModel[];
}